Split configstore.go to more files
